backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
python-telegram-bot>=21.0
Pillow>=10.0
gunicorn>=22.0
python-dotenv>=1.0
